work around GCC -Wextra finding issue with mbiCAUmul overzealously:
we have a test that avoids doing the expensive division if both sides of
the multiplication are smaller than halftype (e.g. 0x00FF for ushort), as
halftype*halftype (e.g. 0x00FF*0x00FF) will fit the full type, no overflow
but GCC in its neverending wisdom decides that that’s worth a warning (the
test will always be true for the second parameter if it’s e.g. int and ut
is longlong due to type size), despite being buried in a generic macro; no
amount of trying (e.g. with ternary, __builtin_choose_expr, …) avoids that
so we test the OR’d value of both operands now, which yields (with MirBSD
GCC for i386) the best code